+++ src/components/MapLayerControlHillshading.tsx
@@ -19,7 +19,7 @@ import { debounce, get } from 'lodash-es';
 /**
  * react-native-mapsforge-vtm dependencies
  */
-import { LayerHillshading, MapContainerProps, ShadingAlgorithm, ShadingAlgorithmOptions } from 'react-native-mapsforge-vtm';
+import { LayerHillshading, ShadingAlgorithm, ShadingAlgorithmOptions } from 'react-native-mapsforge-vtm';
 
 /**
  * Internal dependencies
@@ -32,7 +32,9 @@ import ModalWrapper from './ModalWrapper';
 import { NumericRowControl, NumericMultiRowControl } from './NumericRowControls';
 import ListItemMenuControl from './ListItemMenuControl';
 import HgtSourceRowControl from './HgtSourceRowControl';
-import { fillLayerConfigOptionsWithDefaults } from '../utils';
+import { fillLayerConfigOptionsWithDefaults, getHillshadingCacheDirChild } from '../utils';
+import CacheControl from './CacheControl';
+import { defaults } from '../constants';
 
 const AlgorithmControl = ( {
     options,
@@ -228,13 +230,11 @@ const MapLayerControlHillshading = ( {
             Info={ t( 'hint.maps.shadingMagnitude' ) }
         />
 
-        <NumericRowControl
-            label={ t( 'cacheSize' ) }
-            optKey={ 'cacheSize' }
+        <CacheControl
             options={ options }
             setOptions={ setOptions }
-            validate={ val => val >= 0 }
-            // Info={ t( 'hint.maps.cacheSize' ) }  // ??? is it mb.is the cache working at all ???
+            baseDefault={ defaults.layerConfigOptions.hillshading.cacheDirBase }
+            cacheDirChild={ getHillshadingCacheDirChild( options ) }
         />
 
     </View>;